+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added `showLabel` to `sku-highlights`
 - Indicated unavailability of SKU highlights
 - Added `showLabel` to `sku-inventory`
+- Added number input support for `sku-quantity-selector`
 
 ### Changed
 - Changed SKU image carousel to new version